requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
python-dotenv
passlib[bcrypt]
jwt
pillow
stripe
sendgrid
